==> 1.industry26/4-1.Similarity Multi and Single label.ipynb <==
from ArticlesRep import MeanSimilarityoneindustry,MeanSimilaritytwoindustry,MakeMatrixTSEN #common function

import pandas as pd

import numpy as np

%matplotlib inline

import pickle

import visdom

from tqdm import tnrange, tqdm_notebook
from tqdm.autonotebook import tqdm
tqdm.pandas()

from time import time,sleep
from datetime import datetime

from sklearn.metrics.pairwise import cosine_similarity
# import visdom
# vis=visdom.Visdom()
# env="TagBased"

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
# %matplotlib inline

import re
import sys
import os
from pprint import pprint
#coding:utf-8

#solved chinese display in matplotlib
from pylab import mpl
mpl.rcParams["font.family"] = 'DFKai-sb' 
mpl.rcParams['axes.unicode_minus'] = False

list_industry=["水泥","食品飲料","石化","紡織","電機機械","電器電纜","化學工業",
               "建材居家用品","造紙","鋼鐵金屬","車輛相關","科技相關","營建地產","運輸","觀光休閒娛樂",
               "金融相關","百貨通路","公用事業","控股","生技醫療保健","農林漁牧","航天軍工","能源","傳播出版","綜合",
               "傳產其他","其他","金屬礦採選",
              ]
element="氫氦鋰鈹硼碳氮氧氟氖鈉鎂鋁矽磷硫氯氬鉀鈣鈧鈦釩鉻錳鐵鈷鎳銅鋅鎵鍺砷硒溴氪銣鍶銀鎘銦錫銻碲碘氙銫鋇鉑金汞鉈鉛鉍釙氡鍅鐳"

with open("D:3.AutoencoderForArticle/train_dict_collect_industry_50000.p","rb") as f:
    train_dict_collect_industry_50000=pickle.load(f)
with open("D:3.AutoencoderForArticle/test_dict_collect_industry_50000.p","rb") as f:
    test_dict_collect_industry_50000=pickle.load(f)
with open("D:3.AutoencoderForArticle/dict_multi_industry","rb") as f:
    dict_multi_industry=pickle.load(f)
DATA_S=np.load("D:3.AutoencoderForArticle/BOW_binary_re_v02.npy")
DATA_M=np.load("D:3.AutoencoderForArticle/BOW_binary_multi_v01.npy")
with open("D:3.AutoencoderForArticle/DataSet_vip_multi_industry","rb") as f:
    DataSet_vip_multi_industry=pickle.load(f)
with open("D:3.AutoencoderForArticle/DataSet_vip_single_industry","rb") as f:
    DataSet_vip_single_industry=pickle.load(f)

#### Load Model
from keras.models import load_model

def losspassfunction(y_true,y_pred):
    return y_pred
path="Models/Model5_madeformulti/"
Tri_AutoEncoder=load_model(path+"bestmodel.hdf5",custom_objects={"losspassfunction":losspassfunction})
Tri_AutoEncoder.load_weights(path+"regular/weights.3400.hdf5")
encoder=Tri_AutoEncoder.layers[3]

pd.Series({k:len(dict_multi_industry[k]) for k in dict_multi_industry}).sort_values(ascending=False)

array1=encoder.predict(DATA_M[dict_multi_industry["{'航天軍工', '傳產其他', '公用事業'}"]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
    

pd.Series(dict_simi).sort_values(ascending=False)

array1=encoder.predict(DATA_M[dict_multi_industry["{'化學工業', '石化'}"]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
    

pd.Series(dict_simi).sort_values(ascending=False)

array1=encoder.predict(DATA_M[dict_multi_industry["{'能源', '公用事業'}"]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
    

pd.Series(dict_simi).sort_values(ascending=False)

array1=encoder.predict(DATA_M[dict_multi_industry["{'電機機械', '科技相關', '車輛相關'}"]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False)    


array1=encoder.predict(DATA_M[dict_multi_industry["{'航天軍工', '公用事業', '生技醫療保健', '科技相關', '金融相關', '營建地產', '電機機械', '傳產其他'}"]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

dict_multi_industry["{'航天軍工', '公用事業', '生技醫療保健', '科技相關', '金融相關', '營建地產', '電機機械', '傳產其他'}"]

DataSet_vip_multi_industry[DataSet_vip_multi_industry["indusrty_tags"]=={'航天軍工', '公用事業', '生技醫療保健', '科技相關', '金融相關', '營建地產', '電機機械', '傳產其他'}]


array1=encoder.predict(DATA_M[[11832]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

DataSet_vip_multi_industry[DataSet_vip_multi_industry["indusrty_tags"]=={'能源', '電器電纜', '百貨通路', '營建地產'}]

array1=encoder.predict(DATA_M[[18307]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

DataSet_vip_multi_industry[1523:1524]

array1=encoder.predict(DATA_M[[1523]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

#### 多LABEL也能找出來

DataSet_vip_multi_industry[28000:28001]

array1=encoder.predict(DATA_M[[28000]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

DataSet_vip_multi_industry[53000:53001]

array1=encoder.predict(DATA_M[[53000]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

#### 文章內容與農林漁牧無關

DataSet_vip_multi_industry[38000:38001]

array1=encoder.predict(DATA_M[[38000]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

#### 因為被標有玻璃歸類在建材居家
<li>玻璃‧機殼‧鎂鋁合金外殼‧手機外殼‧可成‧MEMS‧貨幣市場

DataSet_vip_multi_industry[DataSet_vip_multi_industry["guid"].apply(lambda x :str(x))=="9b877400-e8b6-4ae5-9010-7879b1318056"]

array1=encoder.predict(DATA_M[[1636]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

#### 3中2

DataSet_vip_multi_industry[15456:15457]

array1=encoder.predict(DATA_M[[15456]])
dict_simi={}
for i in test_dict_collect_industry_50000:
    array2=encoder.predict(DATA_S[test_dict_collect_industry_50000[i]])
    dict_simi[i]=MeanSimilaritytwoindustry(array1,array2)
pd.Series(dict_simi).sort_values(ascending=False) 

## General

li=[]
dict_multi={k:encoder.predict(DATA_M[dict_multi_industry[k]]) for k in dict_multi_industry}
dict_sing={k:encoder.predict(DATA_S[test_dict_collect_industry_50000[k]]) for k in test_dict_collect_industry_50000}
for k1 in tqdm_notebook(dict_multi):
    array1=dict_multi[k1]
    for k2 in dict_sing:
        array2=dict_sing[k2]
        li.append((k1,k2,MeanSimilaritytwoindustry(array1,array2)))

df=pd.DataFrame(li)
df_pivot=df.pivot(index=0,columns=1,values=2)

df_pivot.iloc[5].sort_values(ascending=False)

dict_AP=dict()
for i,r in tqdm_notebook(df_pivot.iterrows()):
    i_=eval(i)
    i_=i_-set(('金屬礦採選',))
    index=np.asarray(r.sort_values(ascending=False).index)
    mask=np.isin(index,list(i_))
    AP=((np.arange(len(i_))+1)/(np.where(mask)[0]+1)).mean()
    dict_AP[i]=AP

df_AP=pd.Series(dict_AP).sort_values(ascending=False)

#### MAP

print("MAP:",df_AP.mean())
