==> 1.industry26/5.Triplet_DataSet_for_DAE_with_supervision.ipynb <==
from ArticlesRep import MeanSimilarityoneindustry,MeanSimilaritytwoindustry #common function

import pandas as pd

import numpy as np

%matplotlib inline

import pickle

import visdom

from tqdm import tnrange, tqdm_notebook
from tqdm.autonotebook import tqdm
tqdm.pandas()

from time import time,sleep
from datetime import datetime

from sklearn.metrics.pairwise import cosine_similarity
# import visdom
# vis=visdom.Visdom()
# env="TagBased"

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
# %matplotlib inline

import re

import sys
from pprint import pprint
#coding:utf-8
import matplotlib.pyplot as plt
plt.rcParams['font.sans-serif']=['SimHei'] #用来正常显示中文标签
plt.rcParams['axes.unicode_minus']=False #用来正常显示负号
#有中文出现的情况，需要u'内容'
list_industry=["水泥","食品飲料","石化","紡織","電機機械","電器電纜","化學工業",
               "建材居家用品","造紙","鋼鐵金屬","車輛相關","科技相關","營建地產","運輸","觀光休閒娛樂",
               "金融相關","百貨通路","公用事業","控股","生技醫療保健","農林漁牧","航天軍工","能源","傳播出版","綜合",
               "傳產其他","其他","金屬礦採選",
              ]
small_list_insustry=["車輛相關","生技醫療保健","營建地產","百貨通路","傳播出版"]

picksample=np.load("D:3.AutoencoderForArticle/picksample.npy")
len(picksample)

bow=np.load("D:3.AutoencoderForArticle/BOW_binary_v01.npy")

with open("D:3.AutoencoderForArticle/dict_collect_industry.p","rb") as f:
    dict_collect_industry=pickle.load(f)

dict_={k:len(dict_collect_industry[k]) for k in dict_collect_industry}
pd.Series(dict_).sort_values(ascending=False)

train_test_dict_collect_industry=dict()
for k in dict_collect_industry:
    split=int(len(dict_collect_industry[k])*0.8)
    train_test_dict_collect_industry["train_{}".format(k)]=dict_collect_industry[k][:split]
    train_test_dict_collect_industry["test_{}".format(k)]=dict_collect_industry[k][split:] 

dict_={k:len(train_test_dict_collect_industry[k]) for k in train_test_dict_collect_industry}
pd.Series(dict_).sort_values(ascending=False)

# #Save train_test_dict_collect_industry
# with open("D:3.AutoencoderForArticle/train_test_dict_collect_industry.p","wb") as f:
#     pickle.dump(file=f,obj=train_test_dict_collect_industry)

small_list_insustry=["車輛相關","生技醫療保健","營建地產","百貨通路","傳播出版"]

small_train_test_dict_collect_industry=dict()
for i in test_insustry:
    small_train_test_dict_collect_industry["train_{}".format(i)]=train_test_dict_collect_industry["train_{}".format(i)]
    small_train_test_dict_collect_industry["test_{}".format(i)]=train_test_dict_collect_industry["test_{}".format(i)]
    

triplet_train_test_dict_collect_industry=dict()
for i in test_insustry:
    triplet_train_test_dict_collect_industry["train_{}".format(i)]=[]
    triplet_train_test_dict_collect_industry["test_{}".format(i)]=[]

negative=dict()
for k in small_train_test_dict_collect_industry:
    print(k[:5])

small_dict_collect_industry={k:dict_collect_industry[k] for k in small_list_insustry}

negative_pool=dict({k:[] for k in small_dict_collect_industry })
for k in small_dict_collect_industry:
    for j in small_dict_collect_industry:
        if k!=j:negative_pool[k]+=small_dict_collect_industry[j]
    if len(negative_pool[k])!=0:negative_pool[k]=np.random.permutation(negative_pool[k])

## Build dict train/test industry:index in 50,000 single industry VIP news  

with open("D:3.AutoencoderForArticle/DataSet_vip_single_industry","rb") as f:
    Dataset=pickle.load(f)

Dataset.info()

#### build dict_collect_industry_50000

np.random.seed(0)
suffleindex=np.random.permutation(len(Dataset))
pick=suffleindex[:50000]
pick

dict_collect_industry_50000={i:[] for i in list_industry }
for i,j in Dataset.loc[pick].iterrows():
    for k in dict_collect_industry_50000:
        if k in j[2]:
            dict_collect_industry_50000[k].append(i)
            break

dict_count={}
for k in dict_collect_industry_50000:
    dict_count[k]=len(dict_collect_industry_50000[k])
df=pd.Series(dict_count)
df.sort_values(ascending=False)

### build train and test

train_dict_collect_industry_50000=dict()
test_dict_collect_industry_50000=dict()
np.random.seed(10)
rate=3/4
for i in dict_collect_industry_50000:
    induslist=np.random.permutation(dict_collect_industry_50000[i])
    if 0!=len(induslist):
        split=int(np.ceil(len(induslist)*rate))
        train_dict_collect_industry_50000[i]=induslist[:split]
        test_dict_collect_industry_50000[i]=induslist[split:]

#### Save train_test_dict_collect_industry

with open("D:3.AutoencoderForArticle/train_dict_collect_industry_50000.p","wb") as f:
    pickle.dump(file=f,obj=train_dict_collect_industry_50000)
with open("D:3.AutoencoderForArticle/test_dict_collect_industry_50000.p","wb") as f:
    pickle.dump(file=f,obj=test_dict_collect_industry_50000)

dict_count={}
dict_idu_list=train_dict_collect_industry_50000
for k in dict_idu_list:
    dict_count[k]=len(dict_idu_list[k])
df=pd.Series(dict_count)
df.sort_values(ascending=False)

dict_count={}
dict_idu_list=test_dict_collect_industry_50000
for k in dict_idu_list:
    dict_count[k]=len(dict_idu_list[k])
df=pd.Series(dict_count)
df.sort_values(ascending=False)

## Build dict train/test industry:index in all single industry VIP news  (140000 and 10000)

with open("D:3.AutoencoderForArticle/DataSet_vip_single_industry","rb") as f:
    Dataset=pickle.load(f)

Dataset.info()

#### build dict_collect_industry

np.random.seed(5)
suffleindex=np.random.permutation(len(Dataset))
pick=suffleindex
pick

dict_collect_industry={i:[] for i in list_industry }
for i,j in Dataset.loc[pick].iterrows():
    for k in dict_collect_industry:
        if k in j[2]:
            dict_collect_industry[k].append(i)
            break

dict_count={}
for k in dict_collect_industry:
    dict_count[k]=len(dict_collect_industry[k])
df=pd.Series(dict_count)
df.sort_values(ascending=False)

#### build train/test_dict_collect_industry

train_dict_collect_industry=dict()
test_dict_collect_industry=dict()
np.random.seed(10)
rate=(154884-10000)/154884
for i in dict_collect_industry:
    induslist=np.random.permutation(dict_collect_industry[i])
    if 0!=len(induslist):
        split=int(np.floor(len(induslist)*rate))
        train_dict_collect_industry[i]=induslist[:split]
        test_dict_collect_industry[i]=induslist[split:]

dict_count={}
dict_idu_list=train_dict_collect_industry
for k in dict_idu_list:
    dict_count[k]=len(dict_idu_list[k])
df=pd.Series(dict_count)
df.sort_values(ascending=False)

dict_count={}
dict_idu_list=test_dict_collect_industry
for k in dict_idu_list:
    dict_count[k]=len(dict_idu_list[k])
df=pd.Series(dict_count)
df.sort_values(ascending=False)

#### Save train_test_dict_collect_industry

with open("D:3.AutoencoderForArticle/train_dict_collect_industry.p","wb") as f:
    pickle.dump(file=f,obj=train_dict_collect_industry)
with open("D:3.AutoencoderForArticle/test_dict_collect_industry.p","wb") as f:
    pickle.dump(file=f,obj=test_dict_collect_industry)
